Cleaning up from last commit
I had changed the color of the alpha mask to red and gray, I've changed it back to white and gray now.
Fix the window title of text, so it displays.
I think possibly the name of the character should be displayed, rather than just what it looks like, which cannot always be displayed.


git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@696 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2