Various editor changes
MapView "Fill Area" became "Fill Rectanglar Area"
And added a "Fill Area" which behaves like ImageView's "Fill Area" but for tiles.

Made clicking on the spriteset images palette only change
the spriteset indexes if you were viewing the Frames.
(This is so you can safely use the spriteset images palette
as a shortcut when in edit or base and not accidentally
change the order of the sprite, plus changing the order of the indexes
while you couldn't see the changes happening is bound to be annoying.)

Added the ability to reorder the tileset by moving tiles back or forward :)

Added Image->Rotate (so you can freely rotate images in the editor)

Made Image->Rotate, Image->Resize, Image->Rescale, Image->Rescale undoable


git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@751 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2