Making wxeditor better
(Okay, the linux port anyway.)

I added a File->Close because you need it in linux
(It doesn't work correctly though, it just closes the first open document...
Sucks but it atleast allows you to close things in linux.)

Fixed a linux crash that makes certain maps crash.
(It's to do with the tileset scrolling, so I just #ifdef WIN32'ed it out...
It means you cannot scroll anymore, the annoying thing is tilesets that had
enough tiles to require scrolling worked fine, but atleast it doesn't crash.)

Made documents have the correct caption
(i.e. just the filename and not the entire pathname)

I think that's it.


git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@826 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2