Added Script->Tools->Fix Indentation...
(It's not that smart, things like:
if (x)
  y();
or if (x) y();
can't be fixed, only when there's {'s and }'s can it do anything.
Still need to fix indentation found in comment blocks, like:
/**
 *
 */
is found but not properly indented, it wont change it...

Oh well.
Also tried to make the icons register aswell when you register file types...


git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1196 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2